--- src/components/Header/ThemeToggler/ButtonToggle/ButtonToggle.jsx
@@ -1,9 +1,0 @@
-/*
-export const ButtonToggle = ({ children, onClick }) => {
-  return (
-    <button onClick={onClick} type="button">
-      {children}
-    </button>
-  );
-};
-*/
